+++ react/TextAreaBlock.tsx
@@ -48,6 +48,7 @@ const TextAreaBlock: FunctionComponent<TextAreaBlockInterface &
   intl,
   value,
   text,
+  hiddenColumns,
   description,
   componentOnly,
 }: any) => {
@@ -321,6 +322,7 @@ const TextAreaBlock: FunctionComponent<TextAreaBlockInterface &
           <div className={`w-100 ph6 ${handles.reviewBlock}`}>
             <ReviewBlock
               reviewedItems={reviewItems}
+              hiddenColumns={hiddenColumns ?? []}
               onReviewItems={onReviewItems}
               onRefidLoading={onRefidLoading}
             />